Extract colortable imagesource
